requirements.txt
playwright==1.40.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiohttp==3.9.1
python-telegram-bot==20.7
openai==1.3.7
python-dotenv==1.0.0